+++ src/lib.rs
@@ -86,16 +86,19 @@ mod async_subscription;
 #[cfg(feature = "tokio")]
 mod callback;
 mod userdata;
+mod value;
 
 pub use self::{
     client::{Client, ClientBuilder},
     data_type::DataType,
     error::{Error, Result},
     userdata::Userdata,
+    value::{ScalarValue, ValueType, VariantValue},
 };
 pub(crate) use self::{
     data_type::{data_type, enum_variants},
     service::{ServiceRequest, ServiceResponse},
+    value::{ArrayValue, NonScalarValue},
 };
 
 #[cfg(feature = "tokio")]