+++ CHANGELOG.md
@@ -10,12 +10,17 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add `ua::StatusCode::is_uncertain()`, `is_bad()` for checking status code severity.
+- Add support for `ua::Argument` data type and basic support for `ua::ExtensionObject`.
+- Add `Debug` implementation for `ua::Array<T>` data types.
+- Add `ValueType` enum to check `ua::Variant` without unwrapping (also `ua::Argument`).
 
 ### Changed
 
 - Breaking: Return `Result` instead of `Option` for references in `AsyncClient::browse_many()` and
   `browse_next()` (#59).
 - Breaking: Return `Result` instead of raw `ua::DataValue` from `AsyncClient::read_attributes()`.
+- Breaking: Move `ua::VariantValue` and `ua::ScalarValue` to top-level export outside `ua`
+- Breaking: Remove `ua::ArrayValue` for now (until we have a better interface).
 
 ### Fixed
 
@@ -24,6 +29,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `read_attribute()`.
 - Check only severity in `ua::StatusCode::is_good()`. Previously this would be an exact comparison
   to `ua::StatusCode::GOOD`.
+- No longer panic when unwrapping `ua::Variant` with array value.
 
 ## [0.5.0] - 2024-03-01